Feature:  - Switch from GAIA EDR3 to GAIA DR3
Bugfixes:
 - Make code compatible with python 3.12
 - Fixes issues with the newest release of astroquery which returns different keywords for coordinate searches and some photometry catalogs.

 Docs:
  - Update photometry catalogs with GAIA DR3 information.

+++ speedyfit/fileio.py
@@ -101,7 +101,7 @@ def read2array(filename,**kwargs):
    @return: data array (, list of comments)
    @rtype: ndarray (, list)
    """
-   dtype = kwargs.get('dtype',np.float)
+   dtype = kwargs.get('dtype',float)
    return_comments = kwargs.get('return_comments',False)
    data,comm = read2list(filename,**kwargs)
    data = np.array(data,dtype=dtype)
